Add Chacha20Poly1305 and alt cipher support
Use time tracking to set iterations.  Allow the user to set iterations.
Allow users to use Chacha20Poly1305 or AES256-GCM with support for
different ciphers in the future.  Use an AEAD additional data value
to verify that an encrypted file is indeed a LatchBox file.  Update
old LatchBox files to new standard.

+++ src/latchbox/groups.go
@@ -101,7 +101,7 @@ func groupMap() {
       }
     }
     if !inGroupDict {
-      groupDict[group] = string(numToBytes(int64(i), 2))
+      groupDict[group] = string(numToBytes(int(i), 2))
       i++
     }
   }
